SensorManager methods

+++ app/src/main/java/cz/uhk/teeter/Ball.kt
@@ -4,8 +4,8 @@ class Ball {
 
     // velocity, radius, position
     var radius: Int = 0
-    var velocityX: Int = 0
-    var velocityY: Int = 0
+    var velocityX: Float = 0F
+    var velocityY: Float = 0F
     lateinit var position: Point2D
 
 }
